requirements.txt
pandas==2.1.1
requests==2.31.0
streamlit==1.35.0
ipykernel
numpy==1.24.3
scikit-learn==1.3.0
Flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
python-decouple==3.8
